+++ src/main/resources/scripts/recreate-db.sql
@@ -1,22 +1,28 @@
-DROP TABLE users IF EXISTS;
+DROP TABLE users
+IF EXISTS;
+
 CREATE TABLE users (
-  id       INTEGER PRIMARY KEY,
+  id       BIGINT PRIMARY KEY,
   name     VARCHAR(255),
   email    VARCHAR(255),
   birthday VARCHAR(255)
 );
 
-DROP TABLE events IF EXISTS;
+DROP TABLE events
+IF EXISTS;
+
 CREATE TABLE events (
-  id         INTEGER PRIMARY KEY,
+  id         BIGINT PRIMARY KEY,
   name       VARCHAR(255),
   base_price DOUBLE,
   rating     VARCHAR(255)
 );
 
-DROP TABLE tickets IF EXISTS;
+DROP TABLE tickets
+IF EXISTS;
+
 CREATE TABLE tickets (
-  id             INTEGER PRIMARY KEY,
+  id             BIGINT PRIMARY KEY,
   seat           INTEGER,
   is_vip         BOOLEAN,
   date           VARCHAR(255),
